changes in routes and sign out function
