+++ src/Components/Sidebar/Navbar.jsx
@@ -1,16 +1,19 @@
 /* eslint-disable react/prop-types */
-import { Box, Button, Flex } from "@chakra-ui/react";
+import { Box, Button, Flex, useToast } from "@chakra-ui/react";
 import { AiOutlineHome } from "react-icons/ai";
 import { SiExpensify } from "react-icons/si";
 import { PiPiggyBank } from "react-icons/pi";
 import { RiMessage2Line } from "react-icons/ri";
 import { Link } from "react-router-dom";
+import { logout } from "../../Config/firebase";
+import { useState } from "react";
+import { useNavigate } from "react-router-dom";
 
 const navBarItems = [
   {
     iconSrc: <AiOutlineHome />,
     title: "Dashboard",
-    path: "/"
+    path: "/",
   },
   {
     iconSrc: <SiExpensify />,
@@ -28,7 +31,32 @@ const navBarItems = [
   },
 ];
 
-const Navbar = ({ handleLogout, isLoading }) => {
+const Navbar = () => {
+  const [isLoading, setIsLoading] = useState(false);
+
+  const toast = useToast();
+  const navigate = useNavigate();
+
+  const handleLogout = async () => {
+    try {
+      setIsLoading(true);
+      await logout();
+      setIsLoading(false);
+      toast({
+        description: "successfully logged-out",
+        status: "success",
+        duration: 4000,
+        colorScheme: "green",
+      });
+      navigate({ pathname: "/login" });
+    } catch (error) {
+      toast({
+        description: error.message,
+        status: "error",
+        duration: 4000,
+      });
+    }
+  };
   return (
     <>
       <Flex
@@ -41,12 +69,12 @@ const Navbar = ({ handleLogout, isLoading }) => {
         pt={10}
       >
         Dashboard
-        <Box mx={{base: "0", xl: 5}} flex="1" mt={14}>
-          {navBarItems.map((navBarItem, index) => (
+        <Box mx={{ base: "0", xl: 5 }} flex="1" mt={14}>
+          {navBarItems.map((navBarItem) => (
             <Flex
               as={Link}
-              // to={navBarItem.path}
-              key={index}
+              to={navBarItem.path}
+              key={navBarItem.title}
               align="center"
               gap={3}
               _hover={{ backgroundColor: "#ffffff30" }}
