+++ src/Layout/RootLayout.jsx
@@ -1,9 +1,4 @@
-import { Container, Grid, GridItem } from "@chakra-ui/react";
-import Navbar from "../Components/Sidebar/Navbar";
-import Login from "../Pages/Login";
-import Register from "../Pages/Register";
-import Dashboard from "../pages/dashboard/Dashboard";
-import ExpenseBreakdown from "../Components/ExpenseBreakdown";
+import { Container } from "@chakra-ui/react";
 import { Outlet } from "react-router-dom";
 
 const RootLayout = () => {
